fix issue with query param single array value
when a query string is defined as array, but a single param is
passed, req.query will return the param as a primitive type
instead of array. When the value can be coersed to number or
a boolean, a validation error was being triggered, as it wasn't
being recognized as array value.