Add support for registering custom formats
This commit basically makes it where you can pass in a list of custom
formats when creating the `SwaggerApi` object and/or you can call
`SwaggerApi#registerFormat(name, validator)`.  This just proxies through
to `ZSchema#registerFormat(name, validator)` so any value that is
permitted there is permitted here.

This is the bare minimum as there are no features around getting a list
of known formats, being warned if you are re-registering a format, etc.
But since those features were not asked for, this should do what is
required.

Fixes #74