Add support for registering custom format generators
This commit basically makes it where you can pass in a list of custom
format generators when creating the `SwaggerApi` object and/or you can call
`SwaggerApi#registerFormatGenerator(name, formatGenerator)`.

This is the bare minimum as there are no features around getting a list
of known format generators, being warned if you are re-registering a format
generator, etc. But since those features were not asked for, this should do
what is required.

Fixes #100