Fix issue where `UNUSED_DEFINITION` warnings were reported by error
Whenever a referenceable object was created by a relative/remote reference,
references to those objects were not recorded properly and resulted in
unnecessary `UNUSED_DEFINITION` warnings.

This was found when refactoring the tests so that all tests for each file
were ran against the single-file Swagger document and the multi-file
Swagger document.  To make the tests properly work this way, we also had to
add `SwaggerApi#unregisterFormat` and `SwaggerApi#unregisterFormatGenerator`.

Fixes #179